move safeStorage to api
